fix(service): keep the plist reader from resolving markup as a path
The status reads captured-stream paths back out of the installed plist,
and now resolves the escapes the installer writes. A plist mimi did not
write may hold markup rather than escaped text, which parses cleanly and
would resolve to a shorter path naming nothing — so anything that is not
escaped text comes back exactly as it was read.

The escaper and its inverse now sit beside each other, so neither can
drift from the other unnoticed.
