+++ internal/service/plist.go
@@ -2,6 +2,8 @@ package service
 
 import (
 	"encoding/xml"
+	"errors"
+	"io"
 	"path/filepath"
 	"strings"
 )
@@ -131,15 +133,11 @@ func renderPlist(binPath, configPath, logFile, servicePath string) string {
 // escapeXMLText writes it as &#xD;, along with the tab and newline that have
 // the same problem in weaker form.
 //
-// The price is that it also escapes the quote and apostrophe an XML text node
-// does not require escaped. Those render as &#34; and &#39;, which every parser
-// resolves back to the character itself; a path carrying one is rendered more
-// verbosely than it strictly needs, never wrongly.
-//
-// A path that is not valid UTF-8 has no XML representation at all, and
-// encoding/xml substitutes U+FFFD. Such a path cannot reach launchd intact by
-// any escaping; what this buys is a plist launchd still parses, failing at a
-// path that does not exist rather than at a file it will not read.
+// The price is the quote and apostrophe it escapes that a text node does not
+// require escaped, and the U+FFFD it substitutes for a byte sequence that is
+// not valid UTF-8 — which no escaping could have carried into XML anyway. Both
+// cost a path that is written more verbosely, or a path launchd cannot open;
+// neither costs a plist it refuses to read.
 func escapeXMLText(value string) string {
 	var escaped strings.Builder
 
@@ -148,6 +146,55 @@ func escapeXMLText(value string) string {
 	_ = xml.EscapeText(&escaped, []byte(value))
 
 	return escaped.String()
+}
+
+// unescapeXMLText resolves the text of a <string> back to the value it stands
+// for, undoing escapeXMLText for the status that reads an installed plist back.
+// It sits beside its inverse and is encoding/xml doing the work, so that the
+// two sides cannot drift: whatever the escaper learns to write, this reads.
+//
+// Anything that is not escaped text comes back exactly as it was read: text
+// this cannot parse, and equally text it can — markup escapeXMLText would never
+// have written parses cleanly, and resolving it would drop the tags and hand
+// back a shorter path that names nothing. That is a plist mimi did not write —
+// home-manager's, or a hand-edited one — and the reader's bargain everywhere
+// else is to degrade the detail rather than invent one, so the text on disk is
+// the better answer.
+func unescapeXMLText(value string) string {
+	var (
+		decoder = xml.NewDecoder(strings.NewReader("<v>" + value + "</v>"))
+		decoded strings.Builder
+		depth   int
+	)
+
+	for {
+		token, err := decoder.Token()
+		if errors.Is(err, io.EOF) {
+			return decoded.String()
+		}
+
+		if err != nil {
+			return value
+		}
+
+		switch typed := token.(type) {
+		case xml.CharData:
+			decoded.Write(typed)
+		case xml.StartElement:
+			// The wrapper this reads the value inside of is the only element
+			// escaped text can hold.
+			depth++
+			if depth > 1 {
+				return value
+			}
+		case xml.EndElement:
+			depth--
+		default:
+			// A comment, processing instruction or directive is not text
+			// either, and dropping it would shorten the value just as quietly.
+			return value
+		}
+	}
 }
 
 // servicePathFor is the PATH value the plist gets for a given
